revert move service config files
