+++ components/Feed/Replies/ReplyHeader.tsx
@@ -2,9 +2,9 @@ import { deleteDoc, doc } from 'firebase/firestore'
 import { useRouter } from 'next/router'
 import React from 'react'
 
+import { db } from '../../../firebase'
 import { useProfileData } from '../../../hooks/useProfileData'
 import { getAuthorName, getProfilePic } from '../../../lib/profileHelper'
-import { db } from '../../../services/firebase'
 import { postCardClass } from '../../../styles/feed'
 import { authorLabel } from '../../../utils/constants/global'
 import { staticPostData } from '../../../utils/types/params'
