+++ lib/repliesHelper.ts
@@ -9,7 +9,7 @@ import {
     where,
 } from 'firebase/firestore'
 
-import { db } from '../services/firebase'
+import { db } from '../firebase'
 
 /**
  *
