+++ lib/userHelper.ts
@@ -12,7 +12,7 @@ import {
     where,
 } from 'firebase/firestore'
 
-import { db } from '../services/firebase'
+import { db } from '../firebase'
 import { FirebaseProfile, FirebaseUser } from '../utils/types/firebase'
 
 export const getOrCreateUserFromFirebase = async (
