+++ openai.ts
