+++ pages/profile/[id].tsx
@@ -19,9 +19,9 @@ import PostCard from '../../components/Feed/Post/Post'
 import { ProfileCard } from '../../components/Profile/ProfileCard'
 import Button from '../../components/Utils/Button'
 import Modal from '../../components/Utils/Modal'
+import { db } from '../../firebase'
 import useMediaQuery from '../../hooks/useMediaQuery'
 import { useAppSelector } from '../../hooks/useRedux'
-import { db } from '../../services/firebase'
 import { feedApiClass, feedToolbarClass } from '../../styles/feed'
 import { profilePage } from '../../styles/profile'
 import { FirebasePost, FirebaseProfile } from '../../utils/types/firebase'
